OAK-2545 SegmentMk IT tests are too intensive
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1695521 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-core/src/test/java/org/apache/jackrabbit/oak/plugins/segment/SegmentOverflowExceptionIT.java
@@ -21,9 +21,10 @@ package org.apache.jackrabbit.oak.plugins.segment;
 
 import static org.apache.commons.io.FileUtils.deleteDirectory;
 import static org.apache.commons.lang.RandomStringUtils.randomAlphabetic;
-import static org.apache.jackrabbit.oak.plugins.segment.compaction.CompactionStrategy.CleanupType.CLEAN_OLD;
 import static org.apache.jackrabbit.oak.plugins.segment.compaction.CompactionStrategy.MEMORY_THRESHOLD_DEFAULT;
+import static org.apache.jackrabbit.oak.plugins.segment.compaction.CompactionStrategy.CleanupType.CLEAN_OLD;
 import static org.apache.jackrabbit.oak.plugins.segment.file.FileStore.newFileStore;
+import static org.junit.Assume.assumeTrue;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
@@ -33,7 +34,6 @@ import java.util.concurrent.Callable;
 
 import javax.annotation.Nonnull;
 
-import com.google.common.collect.Iterables;
 import org.apache.jackrabbit.oak.api.Blob;
 import org.apache.jackrabbit.oak.api.CommitFailedException;
 import org.apache.jackrabbit.oak.plugins.segment.compaction.CompactionStrategy;
@@ -45,18 +45,29 @@ import org.apache.jackrabbit.oak.spi.state.NodeBuilder;
 import org.apache.jackrabbit.oak.spi.state.NodeStore;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.google.common.collect.Iterables;
+
 /**
- * Test for reproducing OAK-2662. This test will never terminate unless it fails,
- * thus it is marked as @Ignored for now.
+ * <p>Tests verifying if the repository gets corrupted or not: {@code OAK-2662 SegmentOverflowException in HeavyWriteIT on Jenkins}</p>
+ *
+ * <p><b>This test will never terminate unless it fails</b>, thus it is disabled by default. On the
+ * command line specify {@code -DSegmentOverflowExceptionIT=true} to enable
+ * them.</p>
+ *
+ *<p>If you only want to run this test:<br>
+ * {@code mvn verify -Dsurefire.skip.ut=true -PintegrationTesting -Dit.test=SegmentOverflowExceptionIT -DSegmentOverflowExceptionIT=true}
+ * </p>
  */
-@Ignore("long running")
 public class SegmentOverflowExceptionIT {
-    private static final Logger LOG = LoggerFactory.getLogger(SegmentOverflowExceptionIT.class);
+
+    private static final Logger LOG = LoggerFactory
+            .getLogger(SegmentOverflowExceptionIT.class);
+    private static final boolean ENABLED = Boolean
+            .getBoolean(SegmentOverflowExceptionIT.class.getSimpleName());
 
     private final Random rnd = new Random();
 
@@ -64,6 +75,7 @@ public class SegmentOverflowExceptionIT {
 
     @Before
     public void setUp() throws IOException {
+        assumeTrue(ENABLED);
         directory = File.createTempFile(getClass().getSimpleName(), "dir", new File("target"));
         directory.delete();
         directory.mkdir();
@@ -72,7 +84,9 @@ public class SegmentOverflowExceptionIT {
     @After
     public void cleanDir() {
         try {
-            deleteDirectory(directory);
+            if (directory != null) {
+                deleteDirectory(directory);
+            }
         } catch (IOException e) {
             LOG.error("Error cleaning directory", e);
         }
